+++ src/PowerModelsITD.jl
@@ -37,6 +37,12 @@ module PowerModelsITD
     # enables support for v[1]
     Base.getindex(v::JuMP.VariableRef, i::Int) = v
 
+    # const vectors with strings build function names availables.
+    "STANDARD_PROBS constant vector that contains the different types of ITD problems supported."
+    const STANDARD_PROBLEMS = ["build_opfitd", "build_mn_opfitd", "build_pfitd", "build_mn_opfitd_oltc", "build_opfitd_oltc", "build_dmld_opfitd", "build_mn_dmld_opfitd_simple"]
+    "DECOMPOSITION_PROBS constant vector that contains the different types of ITD decomposition problems supported."
+    const DECOMPOSITION_PROBLEMS = ["build_opfitd_decomposition"]
+
     # Files to include in module
     include("io/common.jl")
     include("core/base.jl")